requirements.txt
discord.py
langchain
jinja2
openai
